Add number literal type suffixes (42i64, 3.14f64) and solve Euler 3
Lexer consumes alphanumeric suffixes on number tokens. Type checker
parses and validates suffixes (i32, i64, f64), stripping them before
codegen. Fixed i64 codegen to use std::stoll instead of std::stoi.
Added euler3.mn solving largest prime factor of 600851475143.

+++ src/codegen/CodegenVisitor.cpp
@@ -232,7 +232,7 @@ void CgVisitor::preorder_walk(NumberExprAST *ast) {
     break;
   case TypeKind::I64_t:
     ast->val = llvm::ConstantInt::get(
-        *resPtr->Context, llvm::APInt(64, std::stoi(ast->number), true));
+        *resPtr->Context, llvm::APInt(64, std::stoll(ast->number), true));
     break;
   case TypeKind::F64_t:
     ast->val = llvm::ConstantFP::get(*resPtr->Context,
